[Refactor] 애프터무딩 수정

Closes #233

See merge request s10-fintech-finance-sub2/S10P22C204!199

+++ back/funding/src/main/java/com/ssafy/funding/repository/MovieQueryRepository.java
@@ -58,7 +58,7 @@ public class MovieQueryRepository {
         Expression<Long> likeCountExpression =
                 JPAExpressions.select(movieLike.count())
                         .from(movieLike)
-                        .where(movieLike.movieId.eq(movieId));
+                        .where(movieLike.movieId.eq(movieId.intValue()));
 
         if ("likeAsc".equals(orderByField)) {
             return Expressions.asNumber(likeCountExpression).asc();
